refactor: extract platform-neutral video plumbing into SimUseVideo target
The H.264 Annex B parser, passthrough muxer, AVAssetWriter encoder, frame
utilities, and output-path resolution were platform-neutral but lived in
iOSSimBackend, forcing the Android record-video orchestration into the
SimUse executable target (the only place with both backends in its dep
cone). Move them into a new SimUseVideo target (SimUseCore + system
frameworks only, FB*-free), and sink the generic process-control helpers
(CancellationFlag, OnceFlag, FirstErrorBox, cancellableSleep,
SignalObserver) into SimUseCore.

The single FB*-tied piece — VideoFrameUtilities.captureScreenshotData —
stays in iOSSimBackend as an extension. Pure restructuring, no behavior
change; groundwork for hosting Android video verbs in AndroidBackend (#78).
