+++ Sources/SimUseVideo/AnnexBStreamParser.swift
