+++ Sources/SimUseVideo/VideoOutputFile.swift
@@ -1,0 +1,57 @@
+// SPDX-License-Identifier: Apache-2.0
+import Foundation
+import SimUseCore
+
+/// Output-path resolution shared by every video-producing verb,
+/// regardless of platform backend.
+public enum VideoOutputFile {
+    /// Resolve the user-supplied `--output` argument into a concrete
+    /// MP4 file URL. Both the iOS backend and the cross-platform
+    /// forwarder's Android branch use the same path semantics.
+    public static func prepareOutputURL(output: String?) throws -> URL {
+        let fileManager = FileManager.default
+        let formatter = ISO8601DateFormatter()
+        formatter.formatOptions = [.withInternetDateTime]
+
+        let providedPath = output?.trimmingCharacters(in: .whitespacesAndNewlines)
+        let resolvedPath: String
+        if let providedPath, !providedPath.isEmpty {
+            resolvedPath = (providedPath as NSString).expandingTildeInPath
+        } else {
+            resolvedPath = "sim-use-video-\(formatter.string(from: Date())).mp4"
+        }
+
+        let baseURL: URL
+        if resolvedPath.hasPrefix("/") {
+            baseURL = URL(fileURLWithPath: resolvedPath)
+        } else {
+            baseURL = URL(fileURLWithPath: fileManager.currentDirectoryPath).appendingPathComponent(resolvedPath)
+        }
+
+        var isDirectory: ObjCBool = false
+        if fileManager.fileExists(atPath: baseURL.path, isDirectory: &isDirectory), isDirectory.boolValue {
+            let filename = "sim-use-video-\(formatter.string(from: Date())).mp4"
+            let directoryURL = baseURL
+            if !fileManager.fileExists(atPath: directoryURL.path) {
+                try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
+            }
+            return directoryURL.appendingPathComponent(filename)
+        }
+
+        let directoryURL = baseURL.deletingLastPathComponent()
+        if !fileManager.fileExists(atPath: directoryURL.path) {
+            try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
+        }
+
+        if fileManager.fileExists(atPath: baseURL.path) {
+            var existingIsDirectory: ObjCBool = false
+            fileManager.fileExists(atPath: baseURL.path, isDirectory: &existingIsDirectory)
+            if existingIsDirectory.boolValue {
+                throw CLIError(errorDescription: "Output path \(baseURL.path) is a directory. Provide a file name or point to a different location.")
+            }
+            try fileManager.removeItem(at: baseURL)
+        }
+
+        return baseURL
+    }
+}
