+++ Sources/iOSSimBackend/Verbs/IOSSimRecordVideoCommand.swift
@@ -6,6 +6,7 @@ import FBSimulatorControl
 @preconcurrency import FBControlCore
 import AVFoundation
 import SimUseCore
+import SimUseVideo
 
 /// iOS Simulator backend for the `record-video` verb. Recording uses idb's
 /// native in-process file recorder (`FBSimulator.startRecording(toFile:
@@ -108,7 +109,7 @@ public struct IOSSimRecordVideoCommand: SimUseExecutableCommand {
             throw CLIError(errorDescription: "Simulator \(trimmedUDID) is not booted. Current state: \(stateDescription)")
         }
 
-        let outputURL = try Self.prepareOutputURL(output: output)
+        let outputURL = try VideoOutputFile.prepareOutputURL(output: output)
         FileHandle.standardError.write(Data("Recording simulator \(targetSimulator.udid) to \(outputURL.path)\n".utf8))
         FileHandle.standardError.write(Data("Press Ctrl+C to stop recording\n".utf8))
 
@@ -261,55 +262,5 @@ public struct IOSSimRecordVideoCommand: SimUseExecutableCommand {
         }
 
         try await recorder.finish()
-    }
-
-    /// Resolve the user-supplied `--output` argument into a concrete
-    /// MP4 file URL. Public so the cross-platform forwarder's Android
-    /// branch can reuse the same path semantics.
-    public static func prepareOutputURL(output: String?) throws -> URL {
-        let fileManager = FileManager.default
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime]
-
-        let providedPath = output?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let resolvedPath: String
-        if let providedPath, !providedPath.isEmpty {
-            resolvedPath = (providedPath as NSString).expandingTildeInPath
-        } else {
-            resolvedPath = "sim-use-video-\(formatter.string(from: Date())).mp4"
-        }
-
-        let baseURL: URL
-        if resolvedPath.hasPrefix("/") {
-            baseURL = URL(fileURLWithPath: resolvedPath)
-        } else {
-            baseURL = URL(fileURLWithPath: fileManager.currentDirectoryPath).appendingPathComponent(resolvedPath)
-        }
-
-        var isDirectory: ObjCBool = false
-        if fileManager.fileExists(atPath: baseURL.path, isDirectory: &isDirectory), isDirectory.boolValue {
-            let filename = "sim-use-video-\(formatter.string(from: Date())).mp4"
-            let directoryURL = baseURL
-            if !fileManager.fileExists(atPath: directoryURL.path) {
-                try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
-            }
-            return directoryURL.appendingPathComponent(filename)
-        }
-
-        let directoryURL = baseURL.deletingLastPathComponent()
-        if !fileManager.fileExists(atPath: directoryURL.path) {
-            try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
-        }
-
-        if fileManager.fileExists(atPath: baseURL.path) {
-            var existingIsDirectory: ObjCBool = false
-            fileManager.fileExists(atPath: baseURL.path, isDirectory: &existingIsDirectory)
-            if existingIsDirectory.boolValue {
-                throw CLIError(errorDescription: "Output path \(baseURL.path) is a directory. Provide a file name or point to a different location.")
-            }
-            try fileManager.removeItem(at: baseURL)
-        }
-
-        return baseURL
     }
 }
